Rename model_training to pipeline and removed mredundant model_code
