--- model_training/__init__.py
@@ -1,1 +1,0 @@
-#model_training package
