+++ pyproject.toml
@@ -3,7 +3,7 @@ requires = ["flit_core >=3.2,<4"]
 build-backend = "flit_core.buildapi"
 
 [project]
-name = "model_training"
+name = "pipeline"
 version = "0.0.1"
 description = "training pipeline for REMLA (ass4)"
 authors = [
